% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_upload_file.R
\name{mod_upload_file_ui}
\alias{mod_upload_file_ui}
\alias{mod_upload_file_server}
\title{mod_upload_file_ui and mod_upload_file_server}
\usage{
mod_upload_file_ui(id, label = h4("Dataset "), multiple = TRUE)

mod_upload_file_server(id)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
