% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_read_file_guessing.R
\name{read_file_guessing}
\alias{read_file_guessing}
\title{Read file guessing the extension. Used for Shiny applications.}
\usage{
read_file_guessing(datapath, name, session = session)
}
\arguments{
\item{datapath}{\code{character} specifying the file/s datapath}

\item{name}{\code{character} vector specifying the file/s name}

\item{session}{Used for shiny}
}
\value{
an object of different class depending the function used to read the file
}
\description{
Try to read files guessing the extensions.
  If cant read, an error arise.
}
